fix: 주요 버그 수정 - API 안정성, 금액 오버플로우, 수익률 계산
- [CRITICAL] CompletableFuture.join() → runCatching으로 교체: KRX/VIX/FRED API
  실패 시 fallback 반환, 전체 엔드포인트 500 오류 방지
- [CRITICAL] US Heat 지표 하드코딩(64.0) → VIX + 나스닥/S&P 변동률 기반 동적 계산
- [MEDIUM] Int 오버플로우: currentPrice * quantity (Int*Int) → Long 산술 적용
  (WorkspaceService, WorkspaceEnrichmentService, 도메인 타입, JDBC mapper, SQL V3 마이그레이션)
- [MEDIUM] 모의투자 totalReturnRate: 증분 델타 기준 → 취득 원가 대비 총수익률로 수정

+++ src/main/kotlin/com/giwon/signaldesk/features/market/application/MarketOverviewModel.kt
@@ -203,9 +203,9 @@ data class WatchItem(
 )
 
 data class PortfolioSummary(
-    val totalCost: Int,
-    val totalValue: Int,
-    val totalProfit: Int,
+    val totalCost: Long,
+    val totalValue: Long,
+    val totalProfit: Long,
     val totalProfitRate: Double,
     val positions: List<HoldingPosition>
 )
@@ -217,8 +217,8 @@ data class HoldingPosition(
     val buyPrice: Int,
     val currentPrice: Int,
     val quantity: Int,
-    val profitAmount: Int,
-    val evaluationAmount: Int,
+    val profitAmount: Long,
+    val evaluationAmount: Long,
     val profitRate: Double,
     val source: String = "BASE",
     val id: String = "",
@@ -273,7 +273,7 @@ data class RecommendationExecutionLog(
 
 data class PaperTradingSummary(
     val cash: Int,
-    val evaluation: Int,
+    val evaluation: Long,
     val totalReturnRate: Double,
     val openPositions: List<PaperPosition>,
     val recentTrades: List<PaperTrade>
